+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/analysis/worker/MetricsPersistentWorker.java
@@ -39,6 +39,7 @@ import org.apache.skywalking.oap.server.core.storage.model.Model;
 import org.apache.skywalking.oap.server.core.worker.AbstractWorker;
 import org.apache.skywalking.oap.server.library.client.request.InsertRequest;
 import org.apache.skywalking.oap.server.library.client.request.PrepareRequest;
+import org.apache.skywalking.oap.server.library.client.request.UpdateRequest;
 import org.apache.skywalking.oap.server.library.datacarrier.DataCarrier;
 import org.apache.skywalking.oap.server.library.datacarrier.consumer.BulkConsumePool;
 import org.apache.skywalking.oap.server.library.datacarrier.consumer.ConsumerPoolFactory;
@@ -65,6 +66,11 @@ public class MetricsPersistentWorker extends PersistenceWorker<Metrics> {
      * There are two ways to make sure metrics in-cache,
      * 1. Metrics is read from the Database through {@link #loadFromStorage(List)}
      * 2. The built {@link InsertRequest} executed successfully.
+     *
+     * There are two cases to remove metrics from the cache.
+     * 1. The metrics expired.
+     * 2. The built {@link UpdateRequest} executed failure, which could be caused
+     * (1) Database error. (2) No data updated, such as the counter of update statement is 0 in JDBC.
      */
     private final Map<Metrics, Metrics> sessionCache;
     private final IMetricsDAO metricsDAO;
@@ -72,7 +78,6 @@ public class MetricsPersistentWorker extends PersistenceWorker<Metrics> {
     private final Optional<AbstractWorker<ExportEvent>> nextExportWorker;
     private final DataCarrier<Metrics> dataCarrier;
     private final Optional<MetricsTransWorker> transWorker;
-    private final boolean enableDatabaseSession;
     private final boolean supportUpdate;
     private long sessionTimeout;
     private CounterMetrics aggregationCounter;
@@ -93,7 +98,7 @@ public class MetricsPersistentWorker extends PersistenceWorker<Metrics> {
 
     MetricsPersistentWorker(ModuleDefineHolder moduleDefineHolder, Model model, IMetricsDAO metricsDAO,
                             AbstractWorker<Metrics> nextAlarmWorker, AbstractWorker<ExportEvent> nextExportWorker,
-                            MetricsTransWorker transWorker, boolean enableDatabaseSession, boolean supportUpdate,
+                            MetricsTransWorker transWorker, boolean supportUpdate,
                             long storageSessionTimeout, int metricsDataTTL, MetricStreamKind kind) {
         super(moduleDefineHolder, new ReadWriteSafeCache<>(new MergableBufferedData(), new MergableBufferedData()));
         this.model = model;
@@ -102,7 +107,6 @@ public class MetricsPersistentWorker extends PersistenceWorker<Metrics> {
         // Set to ConcurrentHashMap in order to avoid HashMap deadlock.
         // Since 9.3.0
         this.sessionCache = new ConcurrentHashMap<>(100);
-        this.enableDatabaseSession = enableDatabaseSession;
         this.metricsDAO = metricsDAO;
         this.nextAlarmWorker = Optional.ofNullable(nextAlarmWorker);
         this.nextExportWorker = Optional.ofNullable(nextExportWorker);
@@ -152,14 +156,13 @@ public class MetricsPersistentWorker extends PersistenceWorker<Metrics> {
     MetricsPersistentWorker(ModuleDefineHolder moduleDefineHolder,
                             Model model,
                             IMetricsDAO metricsDAO,
-                            boolean enableDatabaseSession,
                             boolean supportUpdate,
                             long storageSessionTimeout,
                             int metricsDataTTL,
                             MetricStreamKind kind) {
         this(moduleDefineHolder, model, metricsDAO,
              null, null, null,
-             enableDatabaseSession, supportUpdate, storageSessionTimeout, metricsDataTTL, kind
+             supportUpdate, storageSessionTimeout, metricsDataTTL, kind
         );
         // For a down-sampling metrics, we prolong the session timeout for 4 times, nearly 5 minutes.
         // And add offset according to worker creation sequence, to avoid context clear overlap,
@@ -291,7 +294,7 @@ public class MetricsPersistentWorker extends PersistenceWorker<Metrics> {
     }
 
     /**
-     * Load data from the storage, if {@link #enableDatabaseSession} == true, only load data when the id doesn't exist.
+     * Load data from the storage, only load data when the id doesn't exist.
      */
     private void loadFromStorage(List<Metrics> metrics) {
         final long currentTimeMillis = System.currentTimeMillis();
@@ -300,8 +303,8 @@ public class MetricsPersistentWorker extends PersistenceWorker<Metrics> {
                 metrics.stream()
                        .filter(m -> {
                            final Metrics cachedValue = sessionCache.get(m);
-                           // Not cached or session disabled, the metric could be tagged `not in cache`.
-                           if (cachedValue == null || !enableDatabaseSession) {
+                           // the metric is tagged `not in cache`.
+                           if (cachedValue == null) {
                                return true;
                            }
                            // The metric is in the cache, but still we have to check
@@ -326,12 +329,7 @@ public class MetricsPersistentWorker extends PersistenceWorker<Metrics> {
                 return;
             }
 
-            final List<Metrics> dbMetrics = metricsDAO.multiGet(model, notInCacheMetrics);
-            if (!enableDatabaseSession) {
-                // Clear the cache only after results from DB are returned successfully.
-                sessionCache.clear();
-            }
-            dbMetrics.forEach(m -> sessionCache.put(m, m));
+            metricsDAO.multiGet(model, notInCacheMetrics).forEach(m -> sessionCache.put(m, m));
         } catch (final Exception e) {
             log.error("Failed to load metrics for merging", e);
         }
@@ -339,15 +337,13 @@ public class MetricsPersistentWorker extends PersistenceWorker<Metrics> {
 
     @Override
     public void endOfRound() {
-        if (enableDatabaseSession) {
-            Iterator<Metrics> iterator = sessionCache.values().iterator();
-            long timestamp = System.currentTimeMillis();
-            while (iterator.hasNext()) {
-                Metrics metrics = iterator.next();
+        Iterator<Metrics> iterator = sessionCache.values().iterator();
+        long timestamp = System.currentTimeMillis();
+        while (iterator.hasNext()) {
+            Metrics metrics = iterator.next();
 
-                if (metrics.isExpired(timestamp, sessionTimeout)) {
-                    iterator.remove();
-                }
+            if (metrics.isExpired(timestamp, sessionTimeout)) {
+                iterator.remove();
             }
         }
     }